Add-SignUp-and-Login-Logout-ver4
・modelをrun.pyとは別ファイル(models.py)にまとめました
