+++ run.py
@@ -1,53 +1,29 @@
 # Main
 from flask import Flask, render_template, redirect, url_for, flash
-from flask_login import LoginManager, current_user, login_user, logout_user, UserMixin
+from flask_login import LoginManager, current_user, login_user, logout_user
 from flask_bcrypt import Bcrypt
 
-# Form
-from form import ResigtrationForm, LoginForm
+# forms
+from forms import ResigtrationForm, LoginForm
 
+# DB
+from flask_sqlalchemy import SQLAlchemy
 
+# Config
 app = Flask(__name__)
 app.config['SECRET_KEY'] = 'd50afe8ef1fe6f6934245436e6a52776de99f8fa2b31e766991acaf6ef57'
 
 bcrypt = Bcrypt(app)
 login_manager = LoginManager(app)
 
-
-
-# DB
-from flask_sqlalchemy import SQLAlchemy
-from datetime import datetime
-
 db = SQLAlchemy(app)
-# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
 app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql://root:@localhost/flaskblog'
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
 
-@login_manager.user_loader
-def load_user(user_id):
-  return User.query.get(int(user_id))
+# models
+from models import User, Post
 
-class User(db.Model, UserMixin):
-  id = db.Column(db.Integer, primary_key=True)
-  username = db.Column(db.String(20), unique=True, nullable=False)
-  email = db.Column(db.String(120), unique=True, nullable=False)
-  password = db.Column(db.String(60), nullable=False)
-  posts = db.relationship('Post', backref='author', lazy=True)
-
-  def __repr__(self):
-    return f"User('{self.username}', '{self.email}')"
-
-class Post(db.Model):
-  id = db.Column(db.Integer, primary_key=True)
-  title = db.Column(db.String(100), nullable=False)
-  date_posted = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-  content = db.Column(db.Text, nullable=False)
-  user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-
-  def __repr__(self):
-    return f"Post('{self.title}', '{self.date_posted}')"
-
-
+# routes
 @app.route('/')
 def home():
   posts = [
